Split url gen from http req, add timeout fn
The main purpose of this commit is to add timeout fn.

It's now a global option.

Client is created in the run fn and passed to each query executor
(flush, query, test).

query execution is moved out of url builder struct, query execution and
testing and flush are all grouped in main, since they're all IO.